# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Allmax Backend API",
    "version": "1.0.0",
    "description": "API documentation for Allmax Backend"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/llpay/pay": {
      "post": {
        "summary": "Create payment",
        "tags": [
          "LLPay"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Payment parameters depending on payment type"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Database not connected"
          }
        }
      }
    },
...
    "/admin/payments": {
      "get": {
        "summary": "List payments (Admin)",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "txnSeqno",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platformTxno",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}